@@ -1,13 +1,13 @@
 module github.com/openshift/configuration-anomaly-detection
 
-go 1.17
+go 1.19
 
 require (
 	github.com/PagerDuty/go-pagerduty v1.5.0
 	github.com/aws/aws-sdk-go v1.43.2
 	github.com/golang/mock v1.6.0
-	github.com/onsi/ginkgo/v2 v2.1.4
-	github.com/onsi/gomega v1.19.0
+	github.com/onsi/ginkgo/v2 v2.9.1
+	github.com/onsi/gomega v1.27.4
 	github.com/openshift-online/ocm-sdk-go v0.1.299
 	github.com/openshift/aws-account-operator/pkg/apis v0.0.0-20220225132100-6b3881514a02
 	github.com/openshift/hive/apis v0.0.0-20220225023448-bc126c582d02
