@@ -66,7 +66,7 @@ type Client struct {
 }
 
 // New creates a new CCAM client and gets the cluster object from ocm for the internal id
-func New(ocmClient *OcmClient, pdClient *PdClient, externalClusterID string, cluster *v1.Cluster) (Client, error) {
+func New(ocmClient *OcmClient, pdClient *PdClient, cluster *v1.Cluster) (Client, error) {
 	client := Client{
 		Service: Provider{
 			OcmClient: ocmClient,
@@ -87,7 +87,7 @@ func (c Client) checkMissing(err error) bool {
 // Evaluate estimates if the awsError is a cluster credentials are missing error. If it determines that it is,
 // the cluster is placed into limited support, otherwise an error is returned. If the cluster already has a CCAM
 // LS reason, no additional reasons are added and incident is sent to SilentTest.
-func (c Client) Evaluate(awsError error, externalClusterID string, incidentID string) error {
+func (c Client) Evaluate(awsError error, _ string) error {
 	if !c.checkMissing(awsError) {
 		return fmt.Errorf("credentials are there, error is different: %w", awsError)
 	}
